backend_fastapi/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
python-multipart>=0.0.9
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
strands-agents[bidi]>=1.0.0
bedrock-agentcore[strands-agents]
python-dotenv>=1.0.0
pytest>=8.0.0
aiohttp>=3.9.0
